examples/translator: separate out protobufs
Move protobuf definitions out into a separate crate to allow re-use.
