+++ examples/translator/common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "translator_common"
+version = "0.1.0"
+authors = ["David Drysdale <drysdale@google.com>"]
+edition = "2018"
+
+[dependencies]
+oak = "=0.1.0"
+protobuf = "*"
+
+[build-dependencies]
+protoc-rust-grpc = { path = "../../../third_party/grpc-rust/protoc-rust-grpc" }
