+++ examples/translator/common/build.rs
@@ -1,8 +1,8 @@
 fn main() {
     protoc_rust_grpc::run(protoc_rust_grpc::Args {
         out_dir: "src/proto",
-        input: &["../../proto/translator.proto"],
-        includes: &["../../proto", "../../third_party"],
+        input: &["../proto/translator.proto"],
+        includes: &["../proto", "../third_party"],
         rust_protobuf: true, // also generate protobuf messages, not just services
         ..Default::default()
     })
