+++ examples/translator/common/src/proto/mod.rs
