+++ examples/translator/common/src/proto/translator.rs
