+++ examples/translator/common/src/proto/translator_grpc.rs
